Change type of Time from int to float

+++ app/schemas.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel, Field
 
 class Transaction(BaseModel):
     transaction_id: int | None = None
-    Time: int = Field(ge=0)
+    Time: float = Field(ge=0)
     V1: float
     V2: float
     V3: float
